requirements.txt
# Core
python==3.11.*
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ML & Computer Vision
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
pillow==10.2.0
opencv-python==4.9.0.80

# OCR
pytesseract==0.3.10
easyocr==1.7.1

# MLOps
mlflow==2.9.2
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.16.0
apache-airflow-providers-docker==3.9.2

# Data Storage
boto3==1.34.34
minio==7.2.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Monitoring & Logging
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7
loguru==0.7.2

# HTTP & Async
httpx==0.26.0
aiofiles==23.2.1
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tenacity==8.2.3

